final/requirements.txt
numpy
pandas
pydeck
python-dotenv
requests
scikit-learn
spacy
streamlit
